Change the RedfishTransport::Result::body to std::variant<nlohmann::json, bytes> Also change how HTTP transport gets JSON or bytes payloads.
In the previous implementation, payload in the RedfishTransport is assumed to be always JSON. But we need to support non-JSON payload as well.
In this CL, the payload is set to JSON based on default header condition "Content-Type: application/json". But in the future, we will need to set a customized header condition (e.g., "OData-Version: 4") for JSON payload.

PiperOrigin-RevId: 460842069