Add Subscribe RPC
The Subscribe RPC is a server side streaming RPC, as an gRPC transport implementation of the Redfish eventing subsystem.

The RPC takes the existing Request message, such that subscription payload can be set without a separate RPC. Headers are also used for authorization in certain cases.

The RPC streams out the existing Response message. A response can be a JSON formatted Redfish event, or a raw byte stream. Headers in response can be used to indicate the metadata (e.g., subscription ID), and also the format of the response.

PiperOrigin-RevId: 584087641