This CL adds Subscribe RPC abstract interface into both RedfishInterface and the TransportInterface. The Subscribe RPC is a Google's implementation for the Redfish Eventing Subsystem.
The RPC takes a request payload which configures the Event, two callbacks that will be invoked when an event arrives and when the stream closes, and returns an EventStream object. Clients use the EventStream objects to start and stop events.

Currently, only gRPC based transport has Redfish Eventing implemented. The CL leaves the possibility of implementing HTTP SSE style of eventing in the future.

To test the code, eventing on the gRPC fake server is added. The unit test demonstrates the end-to-end workflow. TSAN and ASAN tests indicate no data or thread race.

PiperOrigin-RevId: 587794068