Add virtual `IsStaticInterface()` method to `RedfishInterface`.
This will be used to indicate whether a Redfish interface is static, meaning that it will always return the same payloads from call to call. This can be useful for implementations where the returned `RedfishVariant` payloads are cached by clients, allowing them to dynamically decide to use infinite duration caches in those use cases.

PiperOrigin-RevId: 661442955