Rename `devpath` parameter for `FakeQueryEngine` to more clearly indicate whether Devpath2/Devpath3 is being used.
The current `devpath` variable and `kEnabled`/`kDisabled` does not make it easy to understand which method of devpath resolution is employed in this fake.  `devpath == kEnable` today results in using derived (Devpath2) devpaths and is the default behavior, whereas it might otherwise be assumed this means Devpath3 is the "enabled" state.

PiperOrigin-RevId: 736299325